limited types of resources
limited the types of resources to ones that can be pumped and have a
density. this should exclude electrical stuff and data stuff from other
mods.

i also made it so that it updates available resource types in case you
want to manage new resources from a newly docked vessel.